refactor(tests config): adds exclusions to test coverage and standardizes index files
