+++ sonar-project.properties
@@ -7,15 +7,21 @@ sonar.host.url=https://sonarcloud.io
 
 sonar.sources=./src
 
-# Comma-separated paths to directories with sources (required)
 sonar.language=ts
 
 # TS Metrics
 sonar.typescript.tsconfigPath=./tsconfig.json
 
-sonar.tests=src
-sonar.test.inclusions=**/tests/**
-sonar.exclusions=**/tests/**,**/_app.next.tsx
+sonar.tests=./src
+sonar.test.inclusions=**/*.spec.ts,**/*.spec.tsx,**/tests/**
+sonar.exclusions=
+  **/*.spec.ts, **/*.spec.tsx,**/tests/**,
+  **/_app*.tsx,**/_document*.tsx,
+  **/*.d.ts,
+  **/index.ts,
+  **/index.page.ts,
+  **/styles.ts,
+  **/Theme/**
 
 #----- Default source code encoding
 sonar.sourceEncoding=UTF-8
