+++ src/shared/components/Equalizer/Equalizer.tsx
