+++ src/shared/components/Equalizer/EqualizerSlider/EqualizerSlider.tsx
