+++ src/shared/components/Equalizer/EqualizerSlider/index.ts
@@ -1,0 +1,3 @@
+export { default } from './EqualizerSlider'
+
+export type { EqualizerSliderProps } from './EqualizerSlider';
