+++ src/shared/components/Equalizer/index.ts
@@ -1,0 +1,2 @@
+export { default } from './Equalizer';
+export type { EqualizerProps } from './Equalizer';
