+++ src/shared/components/ReleaseChart/index.ts
@@ -1,0 +1,3 @@
+export { default } from './ReleaseChart';
+
+export type { ReleaseChartProps } from './ReleaseChart';
