+++ src/shared/components/ReleaseValuesDisplay/ReleaseValuesDisplay.tsx
