+++ src/shared/components/ReleaseValuesDisplay/index.ts
@@ -1,0 +1,3 @@
+export { default } from './ReleaseValuesDisplay';
+
+export type { ReleaseValuesDisplayProps } from './ReleaseValuesDisplay';
